requirements.txt
certifi==2020.12.5
chardet==4.0.0
Deprecated==1.2.11
gitdb==4.0.5
GitPython==3.1.14
idna==2.10
PyGithub==1.54.1
PyJWT==1.7.1
PyYAML==5.4.1
requests==2.25.1
smmap==3.0.5
urllib3==1.26.3
wrapt==1.12.1
